SimpleRefractionShader::UpdateMeshData pass: use signed display dimensions
Local: UpdateMeshData 98.9% -> 100% (536/536 bytes).\nGlobal: linked_exact 2585 -> 2586; byte_exact 1160 -> 1160; avg 77.64% -> 77.64%; imports 655 -> 655.

+++ src/engine/render/shaders/SimpleRefractionShader.cpp
@@ -73,8 +73,8 @@ namespace AbyssEngine {
                 glUniform3f(this->uM2Loc, engine->lightColor.x, engine->lightColor.y,
                             engine->lightColor.z);
             int loc = this->uM5Loc;
-            float w = (float) engine->GetDisplayWidth();
-            float h = (float) engine->GetDisplayHeight();
+            float w = (float) (int32_t) engine->GetDisplayWidth();
+            float h = (float) (int32_t) engine->GetDisplayHeight();
             glUniform2f(loc, 1.0f / w, 1.0f / h);
             glActiveTexture(0x84c7);
             engine->ActivateRefractFBO();
